+++ nengo_extras/learning_rules.py
@@ -3,11 +3,120 @@ import warnings
 import numpy as np
 
 from nengo.builder import Builder, Signal
-from nengo.builder.operator import DotInc, ElementwiseInc, Reset, SimPyFunc
+from nengo.builder.connection import get_eval_points, solve_for_decoders
+from nengo.builder.operator import (
+    DotInc, ElementwiseInc, Operator, Reset, SimPyFunc)
 from nengo.exceptions import ValidationError
 from nengo.learning_rules import LearningRuleType
 from nengo.params import EnumParam, FunctionParam, NumberParam
 from nengo.synapses import Lowpass
+
+
+class AML(LearningRuleType):
+    r"""Association matrix learning rule (AML).
+
+    Enables one-shot learning without catastrophic forgetting of outer product
+    association matrices.
+
+    The cue is provided by the pre-synaptic ensemble. The error signal is split
+    up: ``error[0]`` provides a scaling factor to the learning rate.
+    ``error[1]`` provides a decay rate (i.e., weights are multiplied with this
+    value in every time step), ``error[2:]`` provides the target vector.
+
+    The update is given by:
+
+        decoders[...] *= error[1]  # decay
+        decoders[...] += alpha * error[0] * error[2:, None] * np.dot(
+            pre, base_decoders.T)
+
+    where *alpha* is the learning rate adjusted for *dt* and *base_decoders*
+    is the decoder matrix for decoding the identity from the pre-ensemble.
+
+    Parameters
+    ----------
+    d : int
+        Dimensionality of input and output vectors (error signal will be
+        *d+2*).
+    learning_rate : float, optional
+        Learning rate (increase of dot product similarity per second).
+    """
+    error_type = 'decoded'
+    modifies = 'decoders'
+
+    def __init__(self, d, learning_rate=1.):
+        super(AML, self).__init__(learning_rate, size_in=d + 2)
+
+
+class SimAML(Operator):
+    def __init__(self, learning_rate, base_decoders, pre, error, decoders,
+                 tag=None):
+        super(SimAML, self).__init__(tag=tag)
+
+        self.learning_rate = learning_rate
+        self.base_decoders = base_decoders
+
+        self.sets = []
+        self.incs = []
+        self.reads = [pre, error]
+        self.updates = [decoders]
+
+    def make_step(self, signals, dt, rng):
+        base_decoders = self.base_decoders
+        pre = signals[self.pre]
+        error = signals[self.error]
+        decoders = signals[self.decoders]
+        alpha = self.learning_rate * dt
+
+        def step_assoc_learning():
+            scale = error[0]
+            decay = error[1]
+            target = error[2:]
+            decoders[...] *= decay
+            decoders[...] += alpha * scale * target[:, None] * np.dot(
+                pre, base_decoders.T)
+
+        return step_assoc_learning
+
+    @property
+    def pre(self):
+        return self.reads[0]
+
+    @property
+    def error(self):
+        return self.reads[1]
+
+    @property
+    def decoders(self):
+        return self.updates[0]
+
+
+@Builder.register(AML)
+def build_aml(model, aml, rule):
+    conn = rule.connection
+    rng = np.random.RandomState(model.seeds[conn])
+
+    error = Signal(np.zeros(rule.size_in), name="aml:error")
+    model.add_op(Reset(error))
+    model.sig[rule]['in'] = error
+
+    pre = model.sig[conn.pre_obj]['in']
+    decoders = model.sig[conn]['weights']
+
+    encoders = model.params[conn.pre_obj].encoders
+    gain = model.params[conn.pre_obj].gain
+    bias = model.params[conn.pre_obj].bias
+
+    eval_points = get_eval_points(model, conn, rng)
+    targets = eval_points
+
+    x = np.dot(eval_points, encoders.T)
+
+    wrapped_solver = (model.decoder_cache.wrap_solver(solve_for_decoders)
+                      if model.seeded[conn] else solve_for_decoders)
+    base_decoders, _ = wrapped_solver(conn, gain, bias, x, targets, rng=rng)
+
+    model.add_op(SimAML(
+        aml.learning_rate, base_decoders, pre, error, decoders))
 
 
 class DeltaRuleFunctionParam(FunctionParam):